added getProductById with error DTOs and exception handling.

+++ src/main/java/com/shopeasy/estore/product/ProductService.java
@@ -1,9 +1,11 @@
 package com.shopeasy.estore.product;
 
+import com.shopeasy.estore.security.exception.ProductNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -11,8 +13,15 @@ public class ProductService {
     @Autowired
     private ProductRepository productRepository;
     public List<Product> getProductList(){
-
         return productRepository.findAll();
+    }
+
+    public Optional<Product> getProductById(Long id){
+        if(id != null){
+            return productRepository.findById(id);
+        }else{
+            throw new ProductNotFoundException("Product Not Found.");
+        }
     }
 
     public void saveProduct(Product employee) {
@@ -20,6 +29,10 @@ public class ProductService {
     }
 
     public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
+        if(id != null && productRepository.findById(id) != null) {
+            productRepository.deleteById(id);
+        }else{
+            throw new ProductNotFoundException("Product Not Found or Could not be deleted.");
+        }
     }
 }
